Add AppShortcuts.swift
