+++ TCAT/AppDelegate.swift
@@ -23,27 +23,8 @@ class AppDelegate: UIResponder, UIApplicationDelegate {
     let userDefaults = UserDefaults.standard
     
     func application(_ application: UIApplication, didFinishLaunchingWithOptions launchOptions: [UIApplicationLaunchOptionsKey: Any]?) -> Bool {
-        
         // Update shortcut items
-        var shortcutItems = [UIApplicationShortcutItem]()
-        let favorites = SearchTableViewManager.shared.retrieveRecentPlaces(for: Constants.UserDefaults.favorites)
-        for itemType in favorites {
-            switch itemType {
-            case .busStop(let busStop):
-                let data = NSKeyedArchiver.archivedData(withRootObject: busStop)
-                let placeInfo: [AnyHashable: Any] = ["place": data]
-                let shortcutItem = UIApplicationShortcutItem(type: busStop.name, localizedTitle: busStop.name, localizedSubtitle: nil, icon: UIApplicationShortcutIcon(type: .location), userInfo: placeInfo)
-                shortcutItems.append(shortcutItem)
-            case .placeResult(let placeResult):
-                let data = NSKeyedArchiver.archivedData(withRootObject: placeResult)
-                let placeInfo: [AnyHashable: Any] = ["place": data]
-                let shortcutItem = UIApplicationShortcutItem(type: placeResult.name, localizedTitle: placeResult.name, localizedSubtitle: nil, icon: UIApplicationShortcutIcon(type: .location), userInfo: placeInfo)
-                shortcutItems.append(shortcutItem)
-            case .cornellDestination: break
-            case .seeAllStops: break
-            }
-        }
-        UIApplication.shared.shortcutItems = shortcutItems
+        AppShortcuts.shared.updateShortcutItems()
         
         // Set Up Register, Fabric / Crashlytics (RELEASE)
         #if !DEBUG
@@ -107,7 +88,6 @@ class AppDelegate: UIResponder, UIApplicationDelegate {
     }
     
     func application(_ application: UIApplication, performActionFor shortcutItem: UIApplicationShortcutItem, completionHandler: @escaping (Bool) -> Void) {
-        
         handleShortcut(item: shortcutItem)
     }
 
@@ -141,7 +121,6 @@ class AppDelegate: UIResponder, UIApplicationDelegate {
                 return
             }
             optionsVC.searchTo = destination
-            
             if let navController = window?.rootViewController as? UINavigationController{
                 navController.pushViewController(optionsVC, animated: true)
             }
